Former-commit-id: ef2e3abe44ed8c6d8c9947a93c95f789c8a547ad

+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/exec/DBCStatement.java
@@ -97,7 +97,7 @@ public interface DBCStatement extends DBPObject, DBRBlockingObject, DBPCloseable
      * @return number of row updated
      * @throws DBCException on error
      */
-    int getUpdateRowCount() throws DBCException;
+    long getUpdateRowCount() throws DBCException;
 
     /**
      * Checks whether there are additional results (result set or update count).
